Server side Sheet Print,fix
1) Fixed wrong placeholder
2) You can now print using a default PDF template. I uploaded the
template that we use with our LARP game. You have to play on document
coordinates to print in the right sections (see printPGSheet.php).
The character sheet can be printed only if you're authenticated. If you
are a Master, you can print any character sheet. If you are a player,
you can only print the character sheets of your characters.